refactor passBackHeaders to be a shared function
