+++ server/router.js
@@ -1,11 +1,9 @@
 'use strict';
 
 const express = require('express');
-const omit = require('lodash.omit');
 const partial = require('lodash.partial');
 const JSON3 = require('json3');
 const proxy = require('./proxy-request');
-const cloudflare = require('./cloudflare');
 const contentHelper = require('./content-helper');
 const proxyUtil = require('./proxy_util');
 
@@ -13,19 +11,6 @@ let router = express.Router();
 
 let serveLandingPage = function(res) {
   res.sendfile('index.html', { root: __dirname + '/../public' });
-};
-
-let passBackHeaders = function(incomingHeaders) {
-  // remove those that node should generate
-  let resultHeaders = omit(
-    incomingHeaders,
-    'connection',
-    'content-length',
-    'server',
-    'x-frame-options'
-  );
-
-  return cloudflare.filterHeaders(resultHeaders);
 };
 
 let errorToJson = function(error) {
@@ -54,7 +39,7 @@ let doProxy = function(apiUrl, req, res) {
   let promise = proxy(apiUrl, req);
   promise
     .then(function(response) {
-      const responseHeaders = passBackHeaders(
+      const responseHeaders = contentHelper.passBackHeaders(
         contentHelper.iteratorToObj(response.headers)
       );
       res.set(responseHeaders);
